--- examples/example-react-native/src/types.ts
@@ -1,8 +1,0 @@
-import type { TurboModule } from "react-native";
-
-type AsyncStorageDelegate = (value?: boolean) => void;
-
-export type AsyncStorageTestSupport = TurboModule & {
-  test_setDelegate: (delegate: AsyncStorageDelegate) => void;
-  test_unsetDelegate: (delegate: AsyncStorageDelegate) => void;
-};
